Create design tool
Blake Cannoe
Blake Gibson
Vera in Vegas Design tool: Pseudo Code
Declare four different integer arrays. Each array will be a different slot machine. Each of the array’s elements will be the number after you win quarters for that certain slot machine. Create an integer variable for quarters and initialize it to zero. Ask the user how many quarters do they have. Store that number in the quarter’s variable. Output a selection menu to the user that occurs inside of a while loop. The while loop will run until the users quarter’s total is empty or until the user selects to quit the selection menu. Create a switch with six cases. Each case from 1-4 will contain its own slot machine, ex: case 1 = slot one. Each case from 1-4 will have its own specific count variable. When that count variable is equal to the amount of times the slot machine needs to be played to win the case number will add the specific amount of quarters that slot machine pays out to the total. Case 5 will output the amount of quarters the user has at that moment. Case 6 exits the user out of the selection menu and outputs the total number of times played on the slot machines. Also create a default case that outputs: Invalid Input. Create an if statement that will output when the user is out of quarters and exactly how many they played using the count variable. Create four different public static integers for each different slot machine. (Slot one, slot two, slot three, and slot four). Each slot machine will have a count variable, a quarter’s variable and a play’s variable. Each slot machine will have an if statement so when the certain amount of plays reaches its limit, a certain amount of quarters will be added to the total. Create an enhanced for loop so the array data can be read in each slot machine. Create a money variable that divides the total amount of quarters by four to receive a dollar amount. Store that amount in the money variable. Create an else if statement so when the slot machine plays are under the amount that wins the slot machine the quarters will be set to quarters--. Create an output statement that displays the winning amount of quarters plus the users total dollar value. To do so just use the money variable and the quarters variable. Return quarters.